+++ config/api/item-type-allocated-expense/validation.php
@@ -25,7 +25,8 @@ return [
             'total' => [
                 'required',
                 'string',
-                'regex:/^\d+\.\d{2}$/'
+                'regex:/^\d+\.\d{2}$/',
+                'max:16'
             ],
             'percentage' => [
                 'sometimes',